+++ api/models/recipe.js
@@ -6,7 +6,7 @@ const RecipeSchema = new mongoose.Schema({
   ownerId: { type: mongoose.Types.ObjectId, ref: "User", required: true },
   tags: [{ type: String, required: true }],
   favouritedByOwner: { type: Boolean, required: true },
-  totalTime: { type: String, required: true },
+  totalTime: { type: Number, required: true },
   recipeYield: { type: Number, required: true },
   recipeIngredient: [{ type: String, required: true }],
   recipeInstructions: [{ type: String, required: true }],
