--- api/routes/posts.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const PostsController = require("../controllers/posts");
-
-router.get("/", PostsController.getAllPosts);
-router.post("/", PostsController.createPost);
-
-module.exports = router;
